contract-downloader/requirements.txt
python-dotenv
etherscan-python
polygonscan-python
bscscan-python
beautifulsoup4
